Bug #1154942 power: supply: charger-manager: Modify the charging jeita function
[root cause ] Modify the charging jeita function
[changes ] Modify the charging jeita function
[side effects] no
[self test ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: compile ok
[reviewers ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The charger manager will judge whether it is fast charge from the charger driver,
if it is fast charge, we will set the charging current according to the fast charge
jeita table.

Due to the cm_monitor will call in charger manager probe process, so the jeita table
musb be initialization first, otherwise will generate a null pointer, we move the
cm_notify_type_handle before it.

Change-Id: I832db506fe603fb64cd4e3cae7b240d2e6e4cec5
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@unisoc.com>