Bug #1154942 power: supply: charger-manager: Modify the charging jeita function
[root cause  ] Modify the charging jeita function
[changes     ] Modify the charging jeita function
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When in shutdown charging, the cm_monitor will first run, then the charger manager
receive notify type handle, so the jeita table will use the unknown type, if the battery
temperature is not change, the jeita status will not change, so the jeita current will be
incorrect. Now we set jeita charging current according to battery temperature in notify type
handle to solve it.

Change-Id: I0ed9a1c5be2688f1a549edd616a886884a47e767
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>