Bug #1281025 power: supply: charger-manager: Fix high and low temperature cannot stop charging issue
[root cause  ] Fix high and low temperature cannot stop charging issue
[changes     ] Fix high and low temperature cannot stop charging issue
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5pro
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When the phone temperature exceeds the high and low warning lines，the
charger manager monitor this，will stop charging immediately.

When the above scenario occurs，if we plug out the usb cable, then plug in,
it will cannot stop charging, because the high and low warning flag
'emergency_stop' is not clear, it will prevent start charge, the
'info->charging' flag will be set incorrectly.

Change-Id: I115d53974c3628c8a19ff652d3ba845d0e17feff
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>