Bug #1175226 power: supply: charger-manager: Fix the battery temperature jump more than 3 degrees
[root cause  ] Fix the battery temperature jump more than 3 degrees
[changes     ] Fix the battery temperature jump more than 3 degrees
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: build ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Because the ground point of our battery ntc is not the motherboard ground, so
we need to compensate for the difference between the two ground points.
The clock that uses the battery voltage and the current at the battery
terminal are not the same source, Because of this in high temperature
scenarios where the battery temperature is around 50 degrees, if you play
a game, the battery temperature will be more than three degrees.

Now we make up 10 points by taking 15s for the battery temperature value,
and then taking the average and removing the maximum and minimum values
to make up the strategy, We will use the temperature value in the global
variable.

Change-Id: Iba309a0016528e258e8fdcc7c2245f4e22816b55
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spradtrum.com>