Bug #1175226 power: supply: sc27xx: Fix the Low temperature restart battery percentage display issue
[root cause  ] Fix the Low temperature restart battery percentage display issue
[changes     ] Fix the Low temperature restart battery percentage display issue
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: build ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The problem is that when the phone is in a low-temperature environment, the
percentage of power will be converted once according to the temperature curve
after restarting, and the power will be 0% shutdown after several restarts.

In the low temperature environment, we need to save the percentage of
electricity at normal temperature as the initial percentage of electricity at
next startup.

Change-Id: Ie9a947bbbbf3e9c188756b97e5523c7fcf51a128
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spradtrum.com>